# Improve performance of array_map()