The refcounting and destruction is not necessary because zend_call_function will make a copy anyway. And zend_call_function only returns FAILURE if EG(active) is false in which case array_map shouldn't have been called in the first place.

This also adds a fast path for packed arrays (even with holes). By separating the paths of packed/hashed arrays the hashed array iteration is very very slightly faster as well but not by much.

For this script:
```php
$array = range(1, 10000);

for ($i = 0; $i < 800; $i++) {
    array_map(static function ($item) {
            return $item * 2;
    }, $array);
}
```

On an intel i7-1185G7:
```
Benchmark 1: ./sapi/cli/php map.php
  Time (mean ± σ):     127.5 ms ±   3.2 ms    [User: 124.1 ms, System: 3.2 ms]
  Range (min … max):   125.1 ms … 138.1 ms    23 runs
  
Benchmark 2: ./sapi/cli/php_old map.php
  Time (mean ± σ):     148.6 ms ±   3.3 ms    [User: 144.9 ms, System: 3.5 ms]
  Range (min … max):   145.6 ms … 160.6 ms    20 runs
  
Summary
  ./sapi/cli/php map.php ran
    1.17 ± 0.04 times faster than ./sapi/cli/php_old map.php
```